Add template for schedule
Loop through each day in week and create children denoting each ‘day.’
Within each day, loop through each block and display it. Using
bindings, size the block proportional to its duration, add an input
field if it’s a block name (or simply text if it isn’t) that’s bound to
the classes saved in localStorage), and show/hide block number if it’s
a block. There’s quite a hierarchy; looking for ways to simplify it
down.